move onboarding and user stuff into their own directories, outside of auth directory
